ShadowBinder/requirements.txt
tk
qrcode
Pillow
